pyproject.toml
[project]
name = "attorney-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.201.1",
    "crewai-tools[mcp]>=0.75.0",
    "ddgs>=9.6.0",
    "faiss-cpu>=1.12.0",
    "fastmcp>=2.12.4",
    "fpdf>=1.7.2",
    "huggingface-hub>=0.35.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.16.0",
    "numpy>=2.3.3",
    "pdf2image>=1.17.0",
    "pi-heif>=1.1.1",
    "pypdf>=6.1.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "tavily-python>=0.7.12",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "unstructured[local-inference]>=0.18.15",
    "unstructured-inference>=1.0.5",
    "rank-bm25>=0.2.2",
]
